+++ src/scripts/utils/getAlpha.ts
@@ -1,9 +1,4 @@
-/**
- * @param {number} alpha
- * @returns {(string|number)}
- */
-
-const getAlpha = (alpha) =>
+const getAlpha = (alpha: number) =>
   alpha === 1 ? "" : ` / ${alpha.toFixed(2) === "0.00" ? 0 : alpha.toFixed(2)}`;
 
 export default getAlpha;