+++ src/scripts/utils/getArgValue.ts
@@ -1,0 +1,8 @@
+const getArgValue = (flag: string, value: string) => {
+  const index = process.argv.findIndex((arg) => arg === flag);
+  return index !== -1 && process.argv[index + 1]
+    ? process.argv[index + 1]
+    : value;
+};
+
+export default getArgValue;