+++ src/scripts/utils/getColor.js
@@ -1,8 +1,0 @@
-/**
- * @param {number} color
- * @returns {number}
- */
-
-const getColor = (color) => Math.round(color * 255);
-
-export default getColor;