+++ src/scripts/utils/getOpacity.ts
@@ -1,0 +1,3 @@
+const getOpacity = (opacity: number) => opacity / 100;
+
+export default getOpacity;