+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # @heathmont/moon-ui
 
+## 1.5.2
+
+### Patch Changes
+
+- fix: components generation
+
+## 1.5.1
+
+### Patch Changes
+
+- fix: move css files to styles folder
+
+## 1.5.0
+
+### Minor Changes
+
+- feat: add typescript
+
 ## 1.4.1
 
 ### Patch Changes